style: refactor api serializers and views
