+++ backend/api/serializers/__init__.py
@@ -1,0 +1,5 @@
+from .authSerializers import *  # noqa: W0611
+from .courseSerializers import *  # noqa: W0611
+from .exerciseSerializers import *  # noqa: W0611
+from .sessionSerializers import *  # noqa: W0611
+from .studentGroupSerializers import *  # noqa: W0611
