+++ backend/api/views/__init__.py
@@ -1,2 +1,7 @@
 from .views import *  # noqa: W0611 (ignore unused imports warning)
-from .course import *  # noqa: W0611
+from .authViews import *  # noqa: W0611
+from .courseViews import *  # noqa: W0611
+from .exerciseViews import *  # noqa: W0611
+from .resultsViews import *  # noqa: W0611
+from .sessionViews import *  # noqa: W0611
+from .studentGroupViews import *  # noqa: W0611
